working update employee

+++ Fenicia/Fenicia.Application/UseCases/Employees/Delete/DeleteEmployeeInteractor.cs
@@ -1,6 +1,7 @@
 ﻿using Fenicia.Application.Common.Interfaces;
 using Fenicia.Application.Common.Interfaces.UseCases.Employees;
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Fenicia.Application.UseCases.Employees.Delete
@@ -16,14 +17,15 @@ namespace Fenicia.Application.UseCases.Employees.Delete
 
         public async Task<Guid> Handle(DeleteEmployeeRequest request)
         {
-
             var employee = await _context.Employees.FindAsync(request.Id);
-
             if (employee == null)
-            {
-                //throw new NotFoundException(nameof(TodoItem), request.Id);
-            }
+                return Guid.Empty;
+            
+            var user = _context.Users.Single(e => e.Person == employee);
+            if (user == null)
+                return Guid.Empty;
 
+            _context.Users.Remove(user);
             _context.Employees.Remove(employee);
             await _context.SaveChangesAsync();
